refactor(app): Move file upload logic to upload page
